# Preallocate output buffers to improve encode and decode performance