Length of encoded and decoded data is known, so preallocate String and Vec
output buffers in advance. This speeds up benchmarks by a factor of 3.

For decode, check if the given data equals the empty encoding which
avoids allocating at all by returning an empty Vec.