@lopopolo 2020-03-19
# Performance Changes

## Before, master @  3d0743b

https://github.com/artichoke/bubblebabble/runs/519671936?check_suite_focus=true

```
test decode::benchmark_decode_empty             ... bench:          23 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         320 ns/iter (+/- 14)
test decode::benchmark_decode_vector_pineapple  ... bench:         301 ns/iter (+/- 65)
test encode::benchmark_encode_empty             ... bench:          89 ns/iter (+/- 45)
test encode::benchmark_encode_vector_1234567890 ... bench:         291 ns/iter (+/- 13)
test encode::benchmark_encode_vector_pineapple  ... bench:         234 ns/iter (+/- 6)
```

## After, GH-16 @ 823aba1

https://github.com/artichoke/bubblebabble/runs/520236298?check_suite_focus=true

```
test decode::benchmark_decode_empty             ... bench:           4 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         228 ns/iter (+/- 28)
test decode::benchmark_decode_vector_pineapple  ... bench:         204 ns/iter (+/- 32)
test encode::benchmark_encode_empty             ... bench:          24 ns/iter (+/- 3)
test encode::benchmark_encode_vector_1234567890 ... bench:         124 ns/iter (+/- 11)
test encode::benchmark_encode_vector_pineapple  ... bench:         109 ns/iter (+/- 14)
```